fix performance of large graphs for skip binaries computation
